Rewrote part of molprop.cpp.
Split off experiment.cpp that holds data for experiments and
calculations.

Removed junkyard routines to fetch all data from experiments in
one go and made it per MolPropObservable using a std::map.

TODO: fix testing routines

+++ src/alexandria/tests/allbondedsTest.cpp
@@ -77,13 +77,13 @@ class AllBondedsTest : public gmx::test::CommandLineTestBase
                 ob.addPoint(x);
             }
             real   average, sigma;
-            int    N;
+            size_t N;
             eStats ok = ob.getAverageSigmaN(&average, &sigma, &N);
             if (eStats::OK == ok)
             {
                 checker_.checkReal(average, "average");
                 checker_.checkReal(sigma, "sigma");
-                checker_.checkInteger(N, "N");
+                checker_.checkInteger(static_cast<int>(N), "N");
             }
             else
             {
